requirements.txt
streamlit
pymupdf
python-docx
ollama  
